Retrieve dns zone data source with name only.
- Use resource group name if provided, otherwise, use findZone.
- findZone will get the resource groups in your subscription and then
check the zones in each resource group till it finds the first with
a matching dns zone name.
- Could not use the zones client ListComplete api call because
a Zone does not have any properties related to the Resource Group
it is a part of. We have to loop over the list of resource groups
and then their zones so that we can have both the zone and the resource
group name to return and set on the data source.

Fixes #1178.